+++ src/realsense_internal.cpp
@@ -63,46 +63,38 @@ void burt_rs::RealSense::stopStream() {
 
 // -------------------- Frame methods --------------------
 
-// NativeFrames* copyFrame(rs2::frame frame) {
-//   // Copy into a new buffer
-//   const int length = frame.get_data_size();
-//   const uint16_t* data = frame.get_data();
-//   uint16_t* copy = new uint16_t[length];
-//   for (int i = 0; i < length; i++) {
-//     copy[i] = data[i];
-//   }
-//   // Send the buffer and length
-//   auto result = new NativeFrames;
-//   result->data = copy;
-//   result->length = length;
-//   return result;
-// }
-
 NativeFrames* burt_rs::RealSense::getDepthFrame() {
   // Get the depth and color frames
+  // cout << "[BurtRS] Getting frames... " << endl;
   rs2::frameset frames;
   if (!pipeline.poll_for_frames(&frames)) return nullptr;
+  // cout << "[BurtRS]   Getting depth frame" << endl;
   rs2::depth_frame depth_frame = frames.get_depth_frame();
+  // cout << "[BurtRS]   Getting color frame" << endl;
   rs2::frame colorized_frame = colorizer.colorize(depth_frame);
 
-  // Copy both frames -- TODO: optimize this
+  // Copy both frames -- TODO: optimize this to be a move instead
+  // cout << "[BurtRS]   Initializing copies" << endl;
   int depth_length = depth_frame.get_data_size();
   int colorized_length = colorized_frame.get_data_size();
   if (depth_length == 0 || colorized_length == 0) return nullptr;
-  uint16_t* depth_copy = new uint16_t[depth_length];
+  uint8_t* depth_copy = new uint8_t[depth_length];
   uint8_t* colorized_copy = new uint8_t[colorized_length];
 
   // Copy all the data in the depth frame
-  const uint16_t* depth_data = static_cast<const uint16_t*>(depth_frame.get_data());
+  // cout << "[BurtRS]   Copying depth frame" << endl;
+  const uint8_t* depth_data = static_cast<const uint8_t*>(depth_frame.get_data());
   for (int i = 0; i < depth_length; i++) {
     depth_copy[i] = depth_data[i];
   }
 
   // Copy all the data in the colorized frame
+  // cout << "[BurtRS]   Copying colorized frame" << endl;
   const uint8_t* colorized_data = static_cast<const uint8_t*>(colorized_frame.get_data());
   for (int i = 0; i < colorized_length; i++) {
     colorized_copy[i] = colorized_data[i];
   }
+  // cout << "[BurtRS]   Done!" << endl;
 
   // Return both frames
   return new NativeFrames {
@@ -111,25 +103,7 @@ NativeFrames* burt_rs::RealSense::getDepthFrame() {
     colorized_data: colorized_copy,
     colorized_length: colorized_length,
   };
-
 }
-
-//   cout << "[BurtRS] Colorizing " << depthFrame << endl;
-//   rs2::depth_frame* framePtr = reinterpret_cast<rs2::depth_frame*>(depthFrame->rs_pointer);
-//   cout << "[BurtRS] rs2::framePtr: " << framePtr << endl;
-//   rs2::depth_frame frame = *framePtr;
-//   cout << "[BurtRS] Running colorizer..." << endl;
-//   rs2::frame colorized = colorizer.process(*framePtr);
-//   rs2::frame* colorized2 = new rs2::depth_frame(frame);
-//   colorized2->keep();
-//   cout << "[BurtRS] Result: " << colorized2 << endl;
-//   cout << "[BurtRS] Kept result" << endl;
-//   auto result = new BurtRsFrame;
-//   result->data = colorized2->get_data();
-//   result->length = colorized2->get_data_size();
-//   result->rs_pointer = colorized2;
-//   return result;
-// }
 
 void freeFrame(NativeFrames* frames) {
   delete[] frames->depth_data;
